+++ Assets/Prefabs/Ui/MainMenu.prefab
@@ -60,8 +60,8 @@ Transform:
   m_GameObject: {fileID: 2706929141878562863}
   serializedVersion: 2
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 0.22, y: 0.03, z: 1}
-  m_LocalScale: {x: 0.62, y: 0.62, z: 0.62}
+  m_LocalPosition: {x: 0.0388, y: 0, z: 1}
+  m_LocalScale: {x: 0.620806, y: 0.62, z: 0.62}
   m_ConstrainProportionsScale: 1
   m_Children: []
   m_Father: {fileID: 2160977907144639723}
@@ -107,7 +107,7 @@ SpriteRenderer:
   m_SortingLayerID: 0
   m_SortingLayer: 0
   m_SortingOrder: 0
-  m_Sprite: {fileID: 21300000, guid: 8895750559008e83399b8506f16c9030, type: 3}
+  m_Sprite: {fileID: 21300000, guid: e414a37dbba1bb983b0a347bfaa11e54, type: 3}
   m_Color: {r: 1, g: 1, b: 1, a: 1}
   m_FlipX: 0
   m_FlipY: 0
